+++ sim/core/runic_power.go
@@ -931,7 +931,7 @@ type RuneCostImpl struct {
 func newRuneCost(spell *Spell, options RuneCostOptions) *SpellCost {
 	return &SpellCost{
 		spell:           spell,
-		BaseCost:        float64(NewRuneCost(int16(options.RunicPowerCost), options.BloodRuneCost, options.FrostRuneCost, options.UnholyRuneCost, 0)),
+		BaseCost:        int32(NewRuneCost(int16(options.RunicPowerCost), options.BloodRuneCost, options.FrostRuneCost, options.UnholyRuneCost, 0)),
 		PercentModifier: 100,
 		ResourceCostImpl: &RuneCostImpl{
 			BloodRuneCost:  options.BloodRuneCost,
